fix rpc logic + extract states to separate file
